server/requirements/base.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
Faker==4.17.1
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
whitenoise==5.2.0
